backend/requirements.txt
flask==1.1.2
Flask-SQLAlchemy==2.4.4
python-dotenv==0.14.0
flask-marshmallow==0.13.0
marshmallow-sqlalchemy==0.23.1